make linter happy

+++ examples/spend_scale.go
@@ -7,13 +7,13 @@ import (
 )
 
 func (t *Spend) EncodeScale(enc *scale.Encoder) (total int, err error) {
-	if n, err := scale.EncodeCompact8(enc, t.Type); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact8(enc, uint8(t.Type)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if n, err := t.Body.EncodeScale(enc); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -22,13 +22,13 @@ func (t *Spend) EncodeScale(enc *scale.Encoder) (total int, err error) {
 
 func (t *Spend) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if field, n, err := scale.DecodeCompact8(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.Type = field
+		t.Type = uint8(field)
 	}
 	if n, err := t.Body.DecodeScale(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -37,17 +37,17 @@ func (t *Spend) DecodeScale(dec *scale.Decoder) (total int, err error) {
 
 func (t *SpendBody) EncodeScale(enc *scale.Encoder) (total int, err error) {
 	if n, err := scale.EncodeByteArray(enc, t.Adress[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
-	if n, err := scale.EncodeCompact8(enc, t.Selector); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact8(enc, uint8(t.Selector)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if n, err := t.Payload.EncodeScale(enc); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -56,18 +56,18 @@ func (t *SpendBody) EncodeScale(enc *scale.Encoder) (total int, err error) {
 
 func (t *SpendBody) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if n, err := scale.DecodeByteArray(dec, t.Adress[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if field, n, err := scale.DecodeCompact8(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.Selector = field
+		t.Selector = uint8(field)
 	}
 	if n, err := t.Payload.DecodeScale(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -76,22 +76,22 @@ func (t *SpendBody) DecodeScale(dec *scale.Decoder) (total int, err error) {
 
 func (t *SpendPayload) EncodeScale(enc *scale.Encoder) (total int, err error) {
 	if n, err := t.Arguments.EncodeScale(enc); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if n, err := t.Nonce.EncodeScale(enc); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
-	if n, err := scale.EncodeCompact32(enc, t.GasPrice); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact32(enc, uint32(t.GasPrice)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if n, err := scale.EncodeByteArray(enc, t.Signature[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -100,23 +100,23 @@ func (t *SpendPayload) EncodeScale(enc *scale.Encoder) (total int, err error) {
 
 func (t *SpendPayload) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if n, err := t.Arguments.DecodeScale(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if n, err := t.Nonce.DecodeScale(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if field, n, err := scale.DecodeCompact32(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.GasPrice = field
+		t.GasPrice = uint32(field)
 	}
 	if n, err := scale.DecodeByteArray(dec, t.Signature[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -125,12 +125,12 @@ func (t *SpendPayload) DecodeScale(dec *scale.Decoder) (total int, err error) {
 
 func (t *SpendArguments) EncodeScale(enc *scale.Encoder) (total int, err error) {
 	if n, err := scale.EncodeByteArray(enc, t.Recipient[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
-	if n, err := scale.EncodeCompact64(enc, t.Amount); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact64(enc, uint64(t.Amount)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -139,27 +139,27 @@ func (t *SpendArguments) EncodeScale(enc *scale.Encoder) (total int, err error) 
 
 func (t *SpendArguments) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if n, err := scale.DecodeByteArray(dec, t.Recipient[:]); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
 	}
 	if field, n, err := scale.DecodeCompact64(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.Amount = field
+		t.Amount = uint64(field)
 	}
 	return total, nil
 }
 
 func (t *SpendNonce) EncodeScale(enc *scale.Encoder) (total int, err error) {
-	if n, err := scale.EncodeCompact32(enc, t.Counter); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact32(enc, uint32(t.Counter)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
-	if n, err := scale.EncodeCompact64(enc, t.Bitfield); err != nil {
-		return total, err
+	if n, err := scale.EncodeCompact64(enc, uint64(t.Bitfield)); err != nil {
+		return total, err // nolint
 	} else {
 		total += n
 	}
@@ -168,16 +168,16 @@ func (t *SpendNonce) EncodeScale(enc *scale.Encoder) (total int, err error) {
 
 func (t *SpendNonce) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if field, n, err := scale.DecodeCompact32(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.Counter = field
+		t.Counter = uint32(field)
 	}
 	if field, n, err := scale.DecodeCompact64(dec); err != nil {
-		return total, err
+		return total, err // nolint
 	} else {
 		total += n
-		t.Bitfield = field
+		t.Bitfield = uint64(field)
 	}
 	return total, nil
 }
